some utilities in b5m manager. algorithm changes.

+++ b5m/b5m_matcher.rb
@@ -6,7 +6,15 @@ require 'fileutils'
 
 top_dir = File.dirname(File.expand_path(__FILE__))
 config_file = File.join(top_dir, "config.yml")
-config = YAML::load(File.open(config_file))
+default_config_file = File.join(top_dir, "config.yml.default")
+if File.exist?(config_file)
+  config = YAML::load(File.open(config_file))
+elsif File.exist?(default_config_file)
+  config = YAML::load(File.open(default_config_file))
+else
+  puts 'config file not found'
+  exit(false)
+end
 match_config = config["matcher"]
 Dir.chdir(File.dirname(config_file)) do
   match_config['path_of'].each_pair do |k,v|
@@ -32,24 +40,19 @@ IO.readlines(category_file).each do |c_line|
   category_regex = a[0]
   cid = a[1]
   category_dir = File.join(work_dir, cid)
-  unless File.exist?(category_dir)
-    Dir.mkdir(category_dir)
-    regex_file = File.join(category_dir, "category")
-    ofs = File.open(regex_file, 'w')
-    ofs.puts(category_regex)
-    ofs.close
-    fan_file = File.join(category_dir, "filter_attrib_name")
-    FileUtils.cp(filter_attrib_name,fan_file)
-    #run c++ matcher program
-    puts "start building attribute index for #{cid}"
-    system("#{matcher_program} -A -Y #{synonym} -C #{cma} -S #{train_scd} -K #{category_dir}")
-    puts "start matching for #{cid}"
-    output = File.join(category_dir, "match")
-    FileUtils.rm_rf(output) if File.exist?(output)
-    system("#{matcher_program} -B -Y #{synonym} -C #{cma} -S #{scd} -K #{category_dir} -O #{output}")
-    system("cat #{output} >> #{result_file}")
-  else
-    puts "#{cid} exists, skip"
-  end
+  Dir.mkdir(category_dir) unless File.exist?(category_dir)
+  regex_file = File.join(category_dir, "category")
+  ofs = File.open(regex_file, 'w')
+  ofs.puts(category_regex)
+  ofs.close
+  fan_file = File.join(category_dir, "filter_attrib_name")
+  FileUtils.cp(filter_attrib_name,fan_file)
+  #run c++ matcher program
+  puts "start building attribute index for #{cid}"
+  system("#{matcher_program} -A -Y #{synonym} -C #{cma} -S #{train_scd} -K #{category_dir}")
+  puts "start matching for #{cid}"
+  output = File.join(category_dir, "match")
+  system("#{matcher_program} -B -Y #{synonym} -C #{cma} -S #{scd} -K #{category_dir}")
+  #system("cat #{output} >> #{result_file}")
 end
